pyproject.toml
[project]
name = "bitcoin-automation-project"
version = "0.1.0"
description = ""
authors = [
    {name = "Yaronmd",email = "yaronmord@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pytest (>=8.3.5,<9.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "matplotlib (>=3.10.3,<4.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "playwright (>=1.52.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
